chess_backend/pyproject.toml

[tool.poetry]
name = "chess_backend"
version = "0.1.0"
description = ""
authors = ["Amir Ahmady <amir.ahmady@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dlib = "^19.24.6"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
mediapipe = "^0.10.18"
imageio = {extras = ["ffmpeg"], version = "^2.36.1"}
pillow = "^11.0.0"
python-multipart = "^0.0.19"
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"